Fix badge pages
